common: use encoding-related func to own module
